dashboard update
finally figured out how to read from a basic .csv and render an ugly version of the table to a dashboard card on the quarto site. I think having an interactive, filterable dashboard for something like tool comparisons is much more interesting than a markdown table where you get lost in the scrolling.