Using DB Browser for SQLite
downloaded new software to use a GUI to manage some of the SQLite tables, but I still want to learn how to do it better via Python and R. I created the initial table using the RSQLite package. Right now I'm just writing changes to the database via DB Browser and pushing them to the Github repo - they may not make a lot of sense yet, because I'm still learning.